Finalização & Documentação
Criação e modificação dos itens restantes e adicionado a documentação ao que foi realizado. Ainda deve ser comentado sobre o uso de abstrações nas Entities com o professor.

+++ src/main/java/MODELS/ENTITY/ADMINISTRATION/Classe.java
@@ -9,6 +9,12 @@ import lombok.NoArgsConstructor;
 
 import java.util.List;
 
+/**
+ * Classe model da entidade Classe
+ * @author Gustavo Stinghen
+ * @since 10/03/2025
+ * @see Student
+ */
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
@@ -17,7 +23,7 @@ public class Classe {
 
     @Id
     @GeneratedValue( strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String name;
@@ -39,6 +45,13 @@ public class Classe {
     @Column(nullable = false)
     private boolean active;
 
+    /**
+     * Método para adicionar um aluno a classe
+     * @param student aluno a ser adicionado em formato de {@link Student}
+     * @return um booleano indicando se o aluno foi adicionado. Se verdadeiro, o aluno foi adicionado a classe. Se falso, o aluno nao foi adicionado a classe
+     * O aluno nao pode ser adicionado se ele ja estiver na lista de alunos
+     * @see Student
+     */
     public boolean addStudent(Student student) {
 
         if (this.students.contains(student)) {
@@ -50,6 +63,13 @@ public class Classe {
 
     }
 
+    /**
+     * Método para remover um aluno da classe
+     * @param student aluno a ser removido em formato de {@link Student}
+     * @return um booleano indicando se o aluno foi removido. Se verdadeiro, o aluno foi removido da classe. Se falso, o aluno nao foi removido da classe
+     * O aluno nao pode ser removido se ele nao estiver na lista de alunos
+     * @see Student
+     */
     public boolean removeStudent(Student student) {
 
         if (this.students.contains(student)) {
