+++ src/main/java/MODELS/ENTITY/USERS/Admin.java
@@ -8,6 +8,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Classe model da entidade Admin
+ * É uma subclasse de {@link User}
+ * É um usuario que tem permissao de administrador
+ * @author Gustavo Stinghen
+ * @since 10/03/2025
+ * @see User
+ */
 @Entity
 @AllArgsConstructor
 @Data
@@ -16,7 +24,7 @@ public class Admin implements User {
 
     @Id
     @GeneratedValue ( strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String username;
